+++ extensions/cornerstone/src/index.js
@@ -2,6 +2,7 @@ import init from './init.js';
 import asyncComponent from './asyncComponent.js';
 import commandsModule from './commandsModule.js';
 import toolbarModule from './toolbarModule.js';
+import CornerstoneViewportDownloadForm from './CornerstoneViewportDownloadForm';
 
 const OHIFCornerstoneViewport = asyncComponent(() =>
   import(
@@ -37,3 +38,5 @@ export default {
     return commandsModule({ servicesManager });
   },
 };
+
+export { CornerstoneViewportDownloadForm };